refactor: get rid of a cyclic dependency and clean code
