+++ app/component/ItinerarySummaryListContainer.js
@@ -6,6 +6,7 @@ import inside from 'point-in-polygon';
 import cx from 'classnames';
 import startsWith from 'lodash/startsWith';
 import { matchShape } from 'found';
+import isEqual from 'lodash/isEqual';
 
 import distance from '@digitransit-search-util/digitransit-search-util-distance';
 import Icon from './Icon';
@@ -14,8 +15,7 @@ import { isBrowser } from '../util/browser';
 import { getZones } from '../util/legUtils';
 import CanceledItineraryToggler from './CanceledItineraryToggler';
 import { itineraryHasCancelation } from '../util/alertUtils';
-import { matchQuickOption } from '../util/planParamUtil';
-import { getModes } from '../util/modeUtils';
+import { getCurrentSettings, getDefaultSettings } from '../util/planParamUtil';
 import { ItinerarySummarySubtitle } from './ItinerarySummarySubtitle';
 import RightOffcanvasToggle from './RightOffcanvasToggle';
 
@@ -190,18 +190,10 @@ function ItinerarySummaryListContainer(
       msgId = 'walk-bike-itinerary-3';
     }
   } else {
-    const quickOption = matchQuickOption(context);
-    const currentModes = getModes(context.config);
-    const modesDefault =
-      Object.entries(context.config.transportModes).every(
-        ([mode, modeConfig]) =>
-          currentModes.includes(mode.toUpperCase()) === modeConfig.defaultValue,
-      ) && currentModes.includes('PUBLIC_TRANSPORT');
-
-    const hasChanges =
-      quickOption === 'saved-settings' ||
-      quickOption === 'custom-settings' ||
-      !modesDefault;
+    const hasChanges = !isEqual(
+      getCurrentSettings(config),
+      getDefaultSettings(config),
+    );
     if (hasChanges) {
       msgId = 'no-route-msg-with-changes';
     } else {
