+++ app/component/SummaryPlanContainer.js
@@ -16,9 +16,9 @@ import ItinerarySummaryListContainer from './ItinerarySummaryListContainer';
 import TimeNavigationButtons from './TimeNavigationButtons';
 import TimeStore from '../store/TimeStore';
 import PositionStore from '../store/PositionStore';
-import { otpToLocation } from '../util/otpStrings';
+import { otpToLocation, getIntermediatePlaces } from '../util/otpStrings';
 import { getRoutePath } from '../util/path';
-import { getIntermediatePlaces, replaceQueryParams } from '../util/queryUtils';
+import { replaceQueryParams } from '../util/queryUtils';
 import withBreakpoint from '../util/withBreakpoint';
 import { addAnalyticsEvent } from '../util/analyticsUtils';
 import { preparePlanParams } from '../util/planParamUtil';
