+++ app/translations.js
@@ -170,7 +170,6 @@ const translations = {
     'route-page.description': 'Rute {shortName}, {longName}',
     'route-page.title': 'Rute {shortName}',
     'route-page.title-short': 'Rute',
-    'route-stop-or-keyword': 'Linje eller stoppested',
     routes: 'Ruter',
     run: 'Løbe',
     save: 'Gem',
@@ -587,15 +586,6 @@ const translations = {
     'route-page.description': 'Linie {shortName}, {longName}',
     'route-page.title': 'Route {shortName}',
     'route-page.title-short': 'Linie',
-    'route-prefer-greenways': 'Fahrradwege bevorzugen',
-    'route-prefer-illuminated': 'Beleuchtete Wege bevorzugen',
-    'route-prefer-paved': 'Asphaltierte Strecken bevorzugen',
-    'route-prefer-walking-routes': 'Bevorzuge Fußwege',
-    'route-prefer-winter-maintenance': 'Bevorzuge Routen mit Winterdienst',
-    'route-preferences': 'Routing-Einstellungen',
-    'route-public-transport-with-bicycle': 'ÖPNV mit Fahrrad',
-    'route-saved-settings': 'Einstellungen speichern',
-    'route-stop-or-keyword': 'Linie oder Haltestelle',
     routes: 'Linien',
     'routes-platforms': 'Linien, Steige',
     'routes-tracks': 'Linien, Gleise',
@@ -1110,15 +1100,6 @@ const translations = {
     'route-page.description': 'Route {shortName} - {longName}',
     'route-page.title': 'Route {shortName}',
     'route-page.title-short': 'Route',
-    'route-prefer-greenways': 'Prefer cycleways',
-    'route-prefer-illuminated': 'Prefer illuminated routes',
-    'route-prefer-paved': 'Prefer paved routes',
-    'route-prefer-walking-routes': 'Prefer walking routes',
-    'route-prefer-winter-maintenance': 'Prefer routes with winter maintenance',
-    'route-preferences': 'Route preferences',
-    'route-public-transport-with-bicycle': 'Public transport with bicycle',
-    'route-saved-settings': 'Saved settings',
-    'route-stop-or-keyword': 'About the route or stop',
     routes: 'Routes',
     'routes-platforms': 'Routes and platforms',
     'routes-tracks': 'Routes and tracks',
@@ -1470,8 +1451,6 @@ const translations = {
     'route-page.description': 'Ruta {shortName} - {longName}',
     'route-page.title': 'Ruta {shortName}',
     'route-page.title-short': 'Ruta',
-    'route-saved-settings': 'Configuración propia',
-    'route-stop-or-keyword': 'Sobre la ruta o parada',
     routes: 'Rutas',
     run: 'Correr',
     save: 'Guardar',
@@ -1937,15 +1916,6 @@ const translations = {
     'route-page.description': 'Linja {shortName}, {longName}',
     'route-page.title': 'Linja {shortName}',
     'route-page.title-short': 'Linja',
-    'route-prefer-greenways': 'Suosi kevyen liikenteen reittejä',
-    'route-prefer-illuminated': 'Suosi valaistuja reittejä',
-    'route-prefer-paved': 'Suosi päällystettyjä reittejä',
-    'route-prefer-walking-routes': 'Suosi kävelyreittejä',
-    'route-prefer-winter-maintenance': 'Suosi talvikunnossapidettyjä reittejä',
-    'route-preferences': 'Reittimieltymykset',
-    'route-public-transport-with-bicycle': 'Pyöräillen ja julkisilla',
-    'route-saved-settings': 'Tallennetut asetukset',
-    'route-stop-or-keyword': 'Tietoja linjasta tai pysäkistä',
     routes: 'Linjat',
     'routes-platforms': 'Linjat ja laiturit',
     'routes-tracks': 'Linjat ja raiteet',
@@ -2325,15 +2295,6 @@ const translations = {
     'route-page.description': 'Ligne {shortName}, {longName}',
     'route-page.title': 'Ligne {shortName}',
     'route-page.title-short': 'Ligne',
-    'route-prefer-greenways': 'Préférer les pistes cyclables',
-    'route-prefer-illuminated': 'Préférer les routes éclairées',
-    'route-prefer-paved': 'Préférer les routes pavées',
-    'route-prefer-walking-routes': 'Préférer les rues piétonnes',
-    'route-prefer-winter-maintenance': 'Préférer les routes déneigées',
-    'route-preferences': "Préférences d'itinéraire",
-    'route-public-transport-with-bicycle': 'Vélo et transport public',
-    'route-saved-settings': 'Paramètres personnalisés',
-    'route-stop-or-keyword': 'Ligne, arrêt ou mot-clé',
     routes: 'Lignes',
     run: 'Lancer',
     save: 'Enregistrer',
@@ -2604,7 +2565,6 @@ const translations = {
     'route-page.description': 'Rute {shortName}, {longName}',
     'route-page.title': 'Rute {shortName}',
     'route-page.title-short': 'Rute',
-    'route-stop-or-keyword': 'Linje eller stoppested',
     routes: 'Ruter',
     run: 'Løping',
     save: 'Lagre',
@@ -3034,15 +2994,6 @@ const translations = {
     'route-page.description': 'Ruta {shortName} - {longName}',
     'route-page.title': 'Ruta {shortName}',
     'route-page.title-short': 'Ruta',
-    'route-prefer-greenways': 'Preferați pistele de biciclete',
-    'route-prefer-illuminated': 'Preferați rutele iluminate',
-    'route-prefer-paved': 'Preferați rutele asfaltate (pavate)',
-    'route-prefer-walking-routes': 'Preferați rutele desemnate de mers pe jos',
-    'route-prefer-winter-maintenance': 'Preferați rutele întreținute iarna',
-    'route-preferences': 'Preferințe privind rutele',
-    'route-public-transport-with-bicycle': 'Transport public cu bicicleta',
-    'route-saved-settings': 'Setări salvate',
-    'route-stop-or-keyword': 'Despre rută sau stație',
     routes: 'Rute',
     'routes-platforms': 'Rute, peroane',
     'routes-tracks': 'Rute, linii',
@@ -3557,15 +3508,6 @@ const translations = {
     'route-page.description': 'Linje {shortName} - {longName}',
     'route-page.title': 'Linje {shortName}',
     'route-page.title-short': 'Linje',
-    'route-prefer-greenways': 'Föredra lättrafikleder',
-    'route-prefer-illuminated': 'Föredra upplysta rutter',
-    'route-prefer-paved': 'Föredra belagda rutter',
-    'route-prefer-walking-routes': 'Föredra gångvägar',
-    'route-prefer-winter-maintenance': 'Föredra rutter med vinterunderhåll',
-    'route-preferences': 'Ruttpreferenser',
-    'route-public-transport-with-bicycle': 'Med cykel och kollektivtrafik',
-    'route-saved-settings': 'Sparade inställningar',
-    'route-stop-or-keyword': 'Linje eller hållplats',
     routes: 'Linjer',
     'routes-platforms': 'Linjer och plattformer',
     'routes-tracks': 'Linjer och spår',
