+++ app/util/planParamUtil.js
@@ -2,24 +2,13 @@ import omitBy from 'lodash/omitBy';
 import moment from 'moment';
 import cookie from 'react-cookie';
 
-import pick from 'lodash/pick';
-import isEqual from 'lodash/isEqual';
-import {
-  filterModes,
-  getDefaultModes,
-  getDefaultTransportModes,
-  getModes,
-  getStreetMode,
-  hasBikeRestriction,
-} from './modeUtils';
-import { otpToLocation } from './otpStrings';
-import { getIntermediatePlaces, getQuerySettings } from './queryUtils';
+import { filterModes, getDefaultModes, getModes } from './modeUtils';
+import { otpToLocation, getIntermediatePlaces } from './otpStrings';
 import { getDefaultNetworks } from './citybikes';
 import {
   getCustomizedSettings,
   getRoutingSettings,
 } from '../store/localStorage';
-import { OptimizeType, QuickOptionSetType, StreetMode } from '../constants';
 import { estimateItineraryDistance } from './geo-utils';
 
 /**
@@ -44,10 +33,9 @@ export const getDefaultSettings = config => {
  * @param {*} config the configuration for the software installation
  * @param {*} query the query part of the current url
  */
-export const getCurrentSettings = (config, query) => ({
+export const getCurrentSettings = config => ({
   ...getDefaultSettings(config),
   ...getCustomizedSettings(),
-  ...getQuerySettings(query),
 });
 
 // These values need to be null so if no values for the variables are defined somewhere else,
@@ -466,115 +454,4 @@ export const preparePlanParams = config => (
     showBikeAndPublicItineraries: config.showBikeAndPublicItineraries,
     showBikeAndParkItineraries: config.showBikeAndParkItineraries,
   };
-};
-
-export const getApplicableQuickOptionSets = context => {
-  const { config } = context;
-  const streetMode = getStreetMode(config).toLowerCase();
-  return [
-    QuickOptionSetType.DefaultRoute,
-    ...(config.quickOptions[streetMode]
-      ? config.quickOptions[streetMode].availableOptionSets
-      : []),
-  ];
-};
-
-export const getQuickOptionSets = context => {
-  const { config } = context;
-  const defaultSettings = getDefaultSettings(config);
-  const customizedSettings = getCustomizedSettings();
-
-  delete defaultSettings.modes;
-  delete customizedSettings.modes;
-
-  const quickOptionSets = {
-    [QuickOptionSetType.DefaultRoute]: {
-      ...defaultSettings,
-    },
-    [QuickOptionSetType.LeastElevationChanges]: {
-      ...defaultSettings,
-      optimize: OptimizeType.Triangle,
-      safetyFactor: 0.1,
-      slopeFactor: 0.8,
-      timeFactor: 0.1,
-    },
-    [QuickOptionSetType.LeastTransfers]: {
-      ...defaultSettings,
-      transferPenalty: 5460,
-      walkReluctance: config.defaultOptions.walkReluctance.less,
-    },
-    [QuickOptionSetType.LeastWalking]: {
-      ...defaultSettings,
-      walkBoardCost: config.defaultOptions.walkBoardCost.more,
-      walkReluctance: config.defaultOptions.walkReluctance.least,
-    },
-    'public-transport-with-bicycle': {
-      ...defaultSettings,
-      modes: [
-        StreetMode.Bicycle,
-        ...getDefaultTransportModes(config).filter(
-          mode => !hasBikeRestriction(config, mode),
-        ),
-      ].join(','),
-    },
-    [QuickOptionSetType.PreferWalkingRoutes]: {
-      ...defaultSettings,
-      optimize: OptimizeType.Safe,
-      walkReluctance: config.defaultOptions.walkReluctance.most,
-    },
-    [QuickOptionSetType.PreferGreenways]: {
-      ...defaultSettings,
-      optimize: OptimizeType.Greenways,
-    },
-  };
-
-  if (customizedSettings && Object.keys(customizedSettings).length > 0) {
-    quickOptionSets[QuickOptionSetType.SavedSettings] = {
-      ...defaultSettings,
-      ...customizedSettings,
-    };
-  }
-
-  return pick(quickOptionSets, getApplicableQuickOptionSets(context));
-};
-
-export const matchQuickOption = context => {
-  const {
-    config,
-    match: {
-      location: { query },
-    },
-  } = context;
-
-  // Find out which quick option the user has selected
-  const quickOptionSets = getQuickOptionSets(context);
-  const matchesOptionSet = (optionSetName, settings) => {
-    if (!quickOptionSets[optionSetName]) {
-      return false;
-    }
-    const quickSettings = { ...quickOptionSets[optionSetName] };
-    delete quickSettings.modes;
-
-    return isEqual(quickSettings, settings);
-  };
-
-  const querySettings = getQuerySettings(query);
-  const currentSettings = getCurrentSettings(config, query);
-  delete querySettings.modes;
-  delete currentSettings.modes;
-
-  if (matchesOptionSet(QuickOptionSetType.SavedSettings, currentSettings)) {
-    return (
-      Object.keys(quickOptionSets)
-        .filter(key => key !== QuickOptionSetType.SavedSettings)
-        .find(key => matchesOptionSet(key, querySettings)) ||
-      QuickOptionSetType.SavedSettings
-    );
-  }
-
-  return (
-    Object.keys(quickOptionSets).find(key =>
-      matchesOptionSet(key, currentSettings),
-    ) || 'custom-settings'
-  );
 };
