+++ package.json
@@ -189,7 +189,6 @@
     "react-relay-network-modern-ssr": "1.4.0",
     "react-sortablejs": "2.0.11",
     "react-swipeable-views": "0.13.9",
-    "react-transition-group": "1.2.1",
     "recompose": "0.30.0",
     "redis": "2.8.0",
     "relay-runtime": "10.0.1",
